requirements.txt
cryptography==44.0.0
dropbox==12.0.2
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
paramiko==3.5.0